Alot of changes,
Added script reference in top module

Lan keep in mind that the reference to script.js must be in the top 
module, I have added this in. 

I am currently working on the php validation. The extent of the JS, was 
to just check the 16 digits, and presence of the prefix 4, to validate 
it is a Visa Number. But, at the moment, live instant validation (with a 
realtime update of a logo) is a little difficult and not within out time 
constraints. )